Split compilation so types do it themselves
This structure has advantages in that it's now easier to pass things
around without having to worry about format conversion so much.

This also makes it possible to properly convert atoms, lists and strings
versus when passed to macros, without having to guess whether a returned
object was actually a passed-in atom encoded as an object, or a
macro-built one that just happened to look like how an atom would have
been encoded.

One thing that this does sacrifice is that quoting things now no longer
produces useful results outside macros.  This could proabably be hacked
around, but I haven't bothered.