Change major version to 0
I have no idea why I made it 1.0.0 for the first commit.  I played with
rebase --root for a bit before concluding that I couldn't understand its
behaviour.

The zero version number is intended to indicate a semantic version that
this is definitely not to be considered a stable API at all.