No implicit return
"Eating my own dogfood" here. :)  The philosophy is to stay close to
plain JavaScript (and to allow extension with user macros), so the core
language shouldn't have fancy logic for figuring out what if anything
functions should return.

It's also really confusing to have a function return something when you
totally didn't intend it to, or to have to write "undefined" on its own
at the end just to get it to not return.