Rename function-macro from `lambda` to `function`
This is just more obviously in line with the philosophy of "minimum
magic", since in JavaScript the equivalent keyword is "function".
Especially now that functions no longer automatically return the last
thing if it's an expression.

This also has the advantage of no longer occupying the "lambda" keyword;
"function" is reserved in JS anyway, so can't be used.