Special-case import target macro table
In the confusing case of a macro returning another macro, the inner
macro should be added to the environment that the outer macro was
defined in.  (Else it would have no effect at all, because the scope
it's in ends immediately as the outer macro returns.)

This fixes that, by passing another optional argument to an AST node's
`compile`, determining where defined macros should be imported instead.