Pass macro env to require'd macros on `this`
Fixes #4.

Macros defined in other modules and `require`'d in could previously not
access `gensym`, `multi` and other such functions that were only defined
inside ordinary function bodies.  This change exposes them also to
macros in required modules, by calling it with `this` set to an
appropriate object.

Downside: This creates a disparity between how macros look when in a
separate module and how they look in the same, as they're defined
differently.  Something to fix later.