Same macro env usage in normal non-required code
It makes sense to have the macro environment work the same way in the
same file as when required from other files.  This removes the macro
environment functions from the normal scope of the macro body, making it
necessary to access them from `this`.

The stuff available in the environment is much better available this
way, since users can just `console.log(this)` inside a macro (or use a
debugging tool) to see what's there.  Less magical, more understandable.