Have macro object return to mean atom
…and compile atom construction to that instead.

Works correctly in nested-function situations, unlike the previous
`this.atom` solution, which fixes #5.

This also means macros are less "magical", since there exists an easily
understandable plain-JS datatype for every part of a lisp form (arrays
are lists, strings are strings, objects are atoms; numbers are atoms too
actually).  Previously, atoms had to be generated with a function call.