Validate macros-produced AST
Closes #1, with an explicit decision forbidding invalid variable names,
and the necessary checks (with esvalidate) to enforce it.  This prevents
users from (accidentally or deliberately) relying on the plain escodegen
backend's leniency.

Example:

    (a-b "hi")

could previously compile to valid JavaScript code despite the invalid
`a-b` variable name—

    a-b('hi');

—but its meaning is totally different!

- - -

Turns out a small bit of eslisp was also relying on invalid variable
names: `null` is supposed to be encoded as

    { type : "Literal", value : null, raw : "null" }

and *not* as

    { type : "Identifier", name : "null" }

as was previously done.  Hence the extra conditional branch when
compiling atoms.  They compile to the same code in every case I can
think of, but here's to standards-compliance and "eating my own
medicine".  :)

- - -

I pinned an exact esvalidate version, because the validation here
currently compares against a human-readable error string, and I'm not
sure whether they consider that part of their public API.