More explicit and flexible macro definition
Super exciting commit!

Macros really are just functions that run at compile time.  This change
makes that super obvious by stripping the magic from the `macro`-macro
by making it expect its second argument to evaluate to a function.  That
function is then bound to the name in the first argument.

Simple example:

    (macro plusOne (function (x) (return `(+ x 1))))

Previously, it would have looked like this:

    (macro plusOne (x) (return `(+ x 1)))

The new form enables macro functions to be created by whatever means (as
general as any other function), such as

-   a `require` call, or
-   calls to other macros, or
-   an [immediately-invoked function expression (IIFE)][1]

And whatever anything else.  If it compiles to a function, it's good.

This doesn't yet completely obsolete the `macros`-block, because
although the IIFE enables executing arbitrary code, it's not possible to
return no or many functions.

[1]: https://en.wikipedia.org/wiki/Immediately-invoked_function_expression