Macros hygienic by default (optionally capturing)
Previously macros would automatically capture the surrounding
environment with all its macros, such that those other macros would be
applied if present in its return values.  This was potentially super
confusing in situations where a macro was intended to be defined as a
separate module, perhaps written by someone else and `require`d in.

So now `macro` is split into `macro` (the non-env-capturing default one)
and `capmacro` (an env-capturing alternative for when you know what
you're doing).