Call cli module "cli" (instead of "eslc")
Calling it "eslc" was a bad idea, since that's what the executable just
happens to be called by default.  The implementation file in `src/`
could get confused with the actual executable.  The "cli" name will also
continue to be correct even if the command line grows capabilities that
do not necessarily imply compilation.

Naming is hard!