Allow macros to require from node_modules too
Previously, only modules given a path (relative or absolute) could be
require'd inside macros.  Packages required by name (as installed from
npm for exmaple) would fail.

This commit introduces a fix and a unit test to check for it.

Turns out the old way wasn't working for this part because the
`require.main.require` function doesn't fully work unless called with
its `this`-context bound to the `require.main` object.