Add specific test for invalid between-macros tree
It is allowed, and should continue to be, for the syntax tree to be in a
nonsensical state when a macro has made it so, as long as a macro that
it's then passed to can correct it before it needs to be evaluated.

It enables some fancy interactions between macros.