Make multi's macros visible in multi's other stmts
Fixes #7.

When macros returned multiple statements, any macros those statements
contained weren't in scope for subsequent statements that were
simultaneously returned.  (See the test if this sounds confusing.)

This was due to the simple fact that macros were only being entered in
the macro table representing the scope they were being returned into,
and not into the local scope's table used to compile the multi-returned
statements.

The fix was to refactor to let macros be imported into multiple tables,
and to additionally specify the local scope's table as well when
compiling a macro, in case of multiple-returns that need it.