Allow macro-returned macros to also use macros
This was a confusing bug.

It took a while for me to realise the existing test for a
macro-returning macro wasn't actually testing for things you would
expect such a macro to do; which is to call more macros in its body.

Basically, when a macro is defined, it creates a copy of the current
macro scope (its "flattened macro table"), so that macros defined
afterwards don't affect what macros its insides see.  If that macro
returns another macro, that inner macro in turn makes a copy of the
outer macro's table.

Turns out doing it twice like was causing the macro table of the inner
macro to be effectively empty.  This happened because the inner scope's
table was based on the root of the outer table, and the flattening of
tables for that was done such that the actual macros weren't in the
root.

The outer table looked like this—

    {
      contents : { ... all the macros in here ... },
      parent : {
        contents : {},
        parent : null
      },
    }

—when it should have looked like—

    {
      contents : {},
      parent : {
        contents : { ... all the macros in here ... }
        parent : null
      }
    }

The fix was just to reverse those.