Rework `.`-macro to support computed member exprs
Fixes #13.

As pointed out by @wtabatkins in
<https://github.com/anko/eslisp/issues/13#issuecomment-142759925> , the
difference between the `.` and `get` macros has been confusing: The list
`(. a b)` has previously compiled to `a.b;`. This required a separate
form `(get a b)` to produce `a[b];`.

With these changes, `(. a b)` compiles to `a[b];`, and `a.b;` is
produced by `(. a "b")`.  This lets us discard the `get` macro, so the
`.` macro corresponds to a member expression with full generality.  This
keeps the language core minimal and easier to reason about.

This change does make the common use-case of long member expression
chains awkward to type if using core eslisp.  For example, what used to
be `(. Array prototype slice)` will need quotes around the properties:
`(. Array "prototype" "slice" "call")`.  However, if that is your
use-case, you should already be using macros (e.g.
[eslisp-propertify][1]) to sugar it.

If you preferred the old behaviour, you can still have it by writing the
appropriate macros.

[1]: https://github.com/anko/eslisp-propertify