Strict AST validity checking, at end of compile
AST errors were previously only being checked when a list was compiled
to a macro.  This meant that a whole lot of otherwise illegal cases were
not caught.  An error wouldn't be thrown either, allowing the erroneous
result to pass to stdout.

This was handy when everything was unstable, because errors would be
everywhere.  But with stuff having stabilised more recently, it starts
to make sense to check properly.

The test changes are necessary to pass the more rigorous checks.  Most
of them are pedantic stuff that aren't really relevant to the thing
being tested, like return statements having to be inside functions
always, and such.