Make macros persist in REPL
This introduces a stateful version of the compiler, which holds on to
the compilation environment and can be given further code to compile
with respect to that environment.  This is great for the REPL, because
it means macros will be stored in the environment and persist for the
session.

It even works with transform macros given on the command line.

Closes #17.