Rename function-constructing macros
tl;dr:

 - `function` is now `lambda`.
 - `declarefunction` is now `function`.

Rationale:

These are more idiomatic, considering that "lambda" is the name for an
anonymous function in a whole lot of languages, and "defun" or
"function" is the name for a named one.

Of course, JavaScript has the fun quirk that even function expressions
("anonymous" functions) can be given names.  Either way, the distinction
exists, so so must separate macros.

Thanks to @impinball and @lhorie for suggestions and comments in chat.