Correct macro-require for relative paths
Calls to `require` in macros were resolving relative to the eslc
executable, not relative to the current working directory.  This fixes
that, such that cwd should always be used.

I thought existing tests were catching this, but turns out the test was
creating the temporary files that it used in the eslisp project
directory, which totally misses the point.  The files are now created in
a separate tmp directory, and the eslc command is run in full in that
dir.

Fixes #25.