Remove `eslc` indirection, update LiveScript dependency
It's unlikely at this stage that any consumers are depending on the location
of the binary, and it's not even documented within this repo at all, but
hardcoded throughout the docs tests. The executable indirection is removed
to now point to the corresponding file in `lib`. All the former references to
the old binary now run `node lib/cli.js`.

This will also open up the possibility of making this buildable on Windows
platforms.

Also, the newest minor update of LiveScript is purely additive and bug fixes
except for generators no longer being automatically hushed, which shouldn't
affect this repo.

http://livescript.net/#changelog