Catch macro errors and rethrow with location data
This commit adds diagnostic information to errors thrown by macros,
which should help when debugging: the fact that it was a macro that
errored, its name, and the source location of where it was called.

Previously, it could be very unclear from just a stack trace what macro
failed and which call caused it to fail.