Compile empty lists to null instead of nothing
This breaks API for how empty lists are compiled.  They used to stand
for no code whatsoever, and now compile to a `null` identifier.

Rationale: `()` compiling to nothing (as it has since
2b47f4b43619b2bf72ec8e2b8465af74f320b8ff), is pretty useless, given that
there is already a way to make nothing by writing nothing.

The for loop macro was relying on being able to compile () into nothing,
but the fix was pretty simple.

Resolves #32.