Remove this.multi from macro env (arrays instead)
Making macros return an instance of the specific passed-in
"multipleReturn" class was used to be necessary before the most recent
major changes to the macro return format, because it avoided confusion
with actual code.  (Arrays used to be used as the S-expression AST
format.)

Now such a requirement is unnecessarily strict.  Reducing reliance on
the `this`-environment passed to a macro reduces API surface area and
makes the macro functions easier to test in isolation.