Refactor: Catch+rethrow instead of passing arg i
That `i` parameter standing for the argument index currently being
processed is non-descriptive and clutters up the place.  It's neater to
prepend the "there was a problem with argument #i" type of error text in
a central place, and have the leaf conditionals just throw an error with
the details.

Also, few of these error cases were being tested.