Make macros work also in web browsers
The `Module` stuff that makes macros be able to `require` from relative
paths doesn't work in a web browser context, even when eslisp is
processed by a packer like browserify or whatnot.

Packers sometimes locally define `require` for the purpose of making
Node.js-style modules work on the web, so just in case one of those is
active and might work, we silently fall back to whatever `require` is in
this scope.