Add `macroRequire`: easier loading from .esl files
With reference to
https://github.com/anko/eslisp/issues/60#issuecomment-718213534

It is a common use-case to write eslisp macros in separate files.  These
are only used during compilation, and so it can be awkward to have a
build system in place that compiles them to JavaScript first so you can
compile your other code.

To make this use a little more convenient, this commit introduces a
macro called `macroRequire`, which can be given a path to a .esl file
for macros to be loaded from.

`macroRequire` is strictly less powerful than the existing
`macro`-macro, in that it cannot be given an arbitrary expression from
which to load a macro.  However, while it has always been possible to
emulate `macroRequire` using just `macro`, the required degree of
familiarity with Node.js' internal module API made it difficult.  Hence
this sugar.